+++ demo-app/app/app.component.ts
@@ -1,16 +1,19 @@
-import { Component, ElementRef, ViewChild, AfterViewInit } from '@angular/core';
+import { Component, ElementRef, ViewChild,  OnInit, OnDestroy } from '@angular/core';
+import {Subject} from 'rxjs/Subject';
+import {FormBuilder, FormGroup} from '@angular/forms';
+import {MatDialog, MatDialogRef, MAT_DIALOG_DATA} from '@angular/material';
 
 @Component({
-  selector: 'app-root',
-  templateUrl: './app.component.html',
-  styleUrls: [
-    './app.component.scss'
-  ],
+    selector: 'app-root',
+    templateUrl: './app.component.html',
+    styleUrls: [
+        './app.component.scss'
+    ],
 })
-export class AppComponent implements AfterViewInit {
-  @ViewChild('header')
-  header: ElementRef;
-  headerHeight: number;
+export class AppComponent implements OnInit, OnDestroy {
+    @ViewChild('header')
+    header: ElementRef;
+    headerHeight: number;
 
   navItems = [
     { name: 'Accessibility', route: '/accessibility-component' },
@@ -32,12 +35,14 @@ export class AppComponent implements AfterViewInit {
     { name: 'List', route: '/list-content' },
     { name: 'Loading Container', route: '/loading-container' },
     { name: 'Nav List', route: '/nav-list' },
+    { name: 'Markdown', route: '/markdown' },
     { name: 'Page', route: '/page' },
     { name: 'Radio Button', route: '/radio-button' },
     { name: 'Right Drawer', route: '/right-drawer' },
     { name: 'Selection List', route: '/selection-list' },
     { name: 'Sidenav', route: '/sidenav' },
     { name: 'Text', route: '/text' },
+    {name: 'Doc Site Contributions', route: '/contribution'},
   ];
 
   testNavItems = [
@@ -51,9 +56,43 @@ export class AppComponent implements AfterViewInit {
     { name: 'Resources', route: 'resources'}
   ];
 
-  ngAfterViewInit () {
-    setTimeout(() => {
-      this.headerHeight = this.header.nativeElement.offsetHeight;
-    });
-  }
+    public formBuilder: FormBuilder;
+    private close$ = new Subject<void>();
+    form: FormGroup;
+    animal: string;
+    name: string;
+    constructor (formBuilder: FormBuilder, public dialog: MatDialog) {
+        this.formBuilder = formBuilder;
+    }
+
+    openDialog(): void {
+        let dialogRef = this.dialog.open(SlackBotDialogComponent, {
+            width: '250px',
+            data: { name: this.name, animal: this.animal }
+        });
+
+        dialogRef.afterClosed().subscribe(result => {
+            console.log('The dialog was closed');
+            this.animal = result;
+        });
+    }
+    ngOnInit(): void {
+        this.form = this.formBuilder.group({
+            myDateRange: this.formBuilder.group({
+                startDate: [null],
+                endDate: [null]
+            })
+        })
+    }
+
+    ngOnDestroy(): void {
+        this.close$.next();
+        this.close$.complete();
+    }
 }
+
+@Component({
+    selector: 'slackbot-dialog',
+    templateUrl: './slackbot-dialog.html',
+})
+export class SlackBotDialogComponent {}