+++ demo-app/app/demo/demo-accessibility/demo-accessibility.component.ts
@@ -4,7 +4,6 @@ const NOTES: string = require('raw-loader!./demo-accessibility.md');
 @Component({
     selector: 'demo-accessibility',
     templateUrl: './demo-accessibility.html',
-    styleUrls: ['./demo-accessibility.css'],
 })
 export class DemoAccessibilityComponent {
     notes = NOTES;