+++ demo-app/app/demo/demo-card/demo-card.component.ts
@@ -4,7 +4,6 @@ const NOTES: string = require('raw-loader!./demo-card.md');
 @Component({
     selector: 'demo-card',
     templateUrl: './demo-card.html',
-    styleUrls: ['./demo-card.css'],
 })
 export class DemoCardComponent {
     notes = NOTES;