+++ demo-app/app/demo/demo-center/demo-center.component.ts
@@ -4,7 +4,6 @@ const NOTES: string = require('raw-loader!./demo-center.md');
 @Component({
     selector: 'demo-center',
     templateUrl: './demo-center.html',
-    styleUrls: ['./demo-center.css'],
 })
 export class DemoCenterComponent {
     notes = NOTES;