+++ demo-app/app/demo/demo-divider/demo-divider.component.ts
@@ -4,7 +4,6 @@ const NOTES: string = require('raw-loader!./demo-divider.md');
 @Component({
     selector: 'demo-divider',
     templateUrl: './demo-divider.html',
-    styleUrls: ['./demo-divider.css'],
 })
 export class DemoDividerComponent {
     notes = NOTES;