+++ demo-app/app/demo/demo-empty-state/demo-empty-state-example/demo-empty-state-example.component.ts
@@ -3,7 +3,6 @@ import { Component } from '@angular/core';
 @Component({
     selector: 'demo-empty-state-example',
     templateUrl: './demo-empty-state-example.html',
-    styleUrls: ['./demo-empty-state-example.css'],
 })
 
 export class DemoEmptyStateExampleComponent {