+++ demo-app/app/demo/demo-field/demo-field.component.ts
@@ -4,7 +4,6 @@ const NOTES: string = require('raw-loader!./demo-field.md');
 @Component({
     selector: 'demo-field',
     templateUrl: './demo-field.html',
-    styleUrls: ['./demo-field.css'],
 })
 export class DemoFieldComponent {
     notes = NOTES;