+++ demo-app/app/demo/demo-icon/demo-icon.component.ts
@@ -5,7 +5,6 @@ const NOTES: string = require('raw-loader!./demo-icon.md');
 @Component({
     selector: 'demo-icon',
     templateUrl: './demo-icon.html',
-    styleUrls: ['./demo-icon.css'],
 })
 
 export class DemoIconComponent {