+++ demo-app/app/demo/demo-info-banner/demo-info-banner.component.ts
@@ -4,7 +4,6 @@ const NOTES: string = require('raw-loader!./demo-info-banner.md');
 @Component({
     selector: 'demo-info-banner',
     templateUrl: './demo-info-banner.html',
-    styleUrls: ['./demo-info-banner.css'],
 })
 export class DemoInfoBannerComponent {
     notes = NOTES;