+++ demo-app/app/demo/demo-loading-container/demo-loading-container.component.ts
@@ -4,7 +4,6 @@ const NOTES: string = require('raw-loader!./demo-loading-container.md');
 @Component({
     selector: 'demo-loading-container',
     templateUrl: './demo-loading-container.html',
-    styleUrls: ['./demo-loading-container.css'],
 })
 export class DemoLoadingContainerComponent {
     notes = NOTES;