+++ demo-app/app/demo/demo-page/demo-page.component.ts
@@ -4,7 +4,6 @@ const NOTES: string = require('raw-loader!./demo-page.md');
 @Component({
     selector: 'demo-page',
     templateUrl: './demo-page.html',
-    styleUrls: ['./demo-page.css'],
 })
 export class DemoPageComponent {
     notes = NOTES;