+++ demo-app/app/demo/demo-radio-button/demo-radio-button.component.ts
@@ -2,13 +2,12 @@ import { Component } from '@angular/core';
 const NOTES: string = require('raw-loader!./demo-radio-button.md');
 
 @Component({
-  selector: 'demo-radio-button',
-  templateUrl: './demo-radio-button.html',
-  styleUrls: ['./demo-radio-button.css'],
+    selector: 'demo-radio-button',
+    templateUrl: './demo-radio-button.html',
 })
 export class DemoRadioButtonComponent {
-  notes = NOTES;
-  rowOfRadioButtons: string;
+    notes = NOTES;
+    rowOfRadioButtons: string;
 
     buttons = [
         'I AM A BUTTON',