+++ demo-app/app/demo/demo-right-drawer/demo-right-drawer.component.ts
@@ -4,7 +4,6 @@ const NOTES: string = require('raw-loader!./demo-right-drawer.md');
 @Component({
     selector: 'demo-right-drawer',
     templateUrl: './demo-right-drawer.html',
-    styleUrls: ['./demo-right-drawer.css'],
 })
 export class DemoRightDrawerComponent {
     notes = NOTES;