+++ demo-app/app/demo/demo-selection-list/demo-selection-list.component.ts
@@ -4,7 +4,6 @@ const NOTES: string = require('raw-loader!./demo-selection-list.md');
 @Component({
     selector: 'demo-selection-list',
     templateUrl: './demo-selection-list.html',
-    styleUrls: ['./demo-selection-list.css'],
 })
 export class DemoSelectionListComponent {
     notes = NOTES;